Prevent users from adding colliding paths
Categories and rulesets are indexed by relative paths that are looked up on the classpath. Prevent users from adding colliding paths which would lead to confusing behavior since the first one found on the classpath is the one that would be run by PMD.

Added the ability to add an XML file from the filesystem.