Second part to tsconfig changes
Allow the user to specify a location of tsconfig.json using a command
line flag. Throw an error if a tsconfig.json is found in the current
working directory and the --tsconfig flag is specified. This error is
thrown even if they are the same file.

Attempted to strike a balance of leaking the CLI into the engines. The
engines should be usable from other callers other than the CLI, but the
validation occurs in the engine itself and the user needs to know how to
fix errors. For now, specify the command line flag using the same name
as the key in the Engine Options.

RuleEngine takes a map of options. Currently only the Typescript engine
accepts or uses these options.