Fix issues with Config#isEngineEnabled
Config#isEngineEnabled had a bug and was returning the wrong value.
None of the engine consumers were filtering out disabled engines.

This is in preparation for adding the eslint-lwc engine which will be
enabled by default.

Made ENGINE_NAME vs. NAME consistent across engines.

Updated Config#isEngineEnabled to take an ENGINE object instead of
string. This makes it consistent with other methods.

Put tests within 'it' blocks to ensure 'before' hooks are executed.