Config#isEngineEnabled should use #getConfigValue
Invoke #getConfigValue so that the defaulting path is used when the engine node is not present. Previously this code path didn't result in an invocation of #lookupAndUpdateToDefault which caused an initial lookup for an engine to always return true, even if the engine should be disabled by default.

This change had a bit of fallout since the #isEngineEnabled method now returns a Promise.

Fixed issue with Config#lookupAndUpdateToDefault where a default of 'false' would fall through to the else block that throws an exception.

Fixed issue with typeChecker#booleanCheck that was incorrectly comparing instanceof to an oclif 'boolean' type. instanceof can't be used on the primitive boolean type. Use typeof instead

Removed export of Config#DEFAULT_CONFIG that made it hard to unit test