Fix filter by Category regression
This restores the JavascriptEslintStrategy and TypescriptEslintStrategy configs to their 2.1 values. While these values aren't correct, they are safe.

Using the extends attribute in the config will turn use the default rule settings for the rules that are extended. We must explicitly turn off any rules that are in the extended rule set if we don't want them to show up in the results.

The correct fix for the Javascript, LWC and Typescript engines will go out in another release.

It is a known issue that the eslint-lwc engine won't filter by Category correctly.