@W-8063535@ Severity normalizer
The severity normalizer adds a normalizedSeverity value to the results when used and displays the normalized value alongside the original severity. 

The -n flag determines if the normalized severities should be calculated and included in the output.
The -s flag throws an error when violations are found of a specific (or greater) severity.