Recreate symlinks to work around nvm issue
nvm-windows 1.1.7 has a bug when combined with npm 8.3.1. This can be fixed with either npm 8.3.2 or nvm-windows 1.1.9. More information can be found here https://github.com/npm/cli/issues/4340#issuecomment-1025833090

The orb comes with nvm-windows 1.1.7 and would be a large effort to remove and install.
The node installation does not work enough to upgrade npm from 8.3.1 to a newer version.

This is a temporary change that allows the Windows tests to pass until the orb's nvw-windows version is updated or a newer version of node is released that bundles npm 8.3.2+.

I expect the tests will start failing when one of these two occur.